Corrected FloorArea type from int to double.

+++ src/ETLAthena.Core/Models/S2Model.cs
@@ -7,6 +7,6 @@ namespace ETLAthena.Core.Models
         public string? Address { get; set; }
         public string? Postcode { get; set; }
         public double[]? Coordinates { get; set; }
-        public int? FloorArea { get; set; }
+        public double? FloorArea { get; set; }
     }
 }
